# Reduce the number of heap allocations when computing hash digests