@baarde 2024-06-26
I've rewritten the code to use the specialized `MD5_*`/`SHA*_*` APIs instead of the generic `EVP_MD_*` APIs.

I've also made sure that context and digest buffers get zeroized after use (`EVP_DigestFinal` and `EVP_MD_CTX_free` already did that for context buffers).

One noticeable "hack" is the use of the `@exclusivity` attribute to get a +15% performance improvement. This attribute was proposed in [SE-0176](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0176-enforce-exclusive-access-to-memory.md) and [implemented](https://github.com/swiftlang/swift/commit/5fd941eb677cdc97333deb051c8811a25fd77c70) in Swift 5.7, but is currently undocumented.